work_bot/requirements.txt
pyTelegramBotAPI==4.14.1
python-dotenv==0.20.0
